+++ ztf_viewer/catalogs/conesearch/__init__.py
@@ -3,6 +3,7 @@ from .alerce import AlerceQuery
 from .antares import AntaresQuery
 from .astrocats import AstrocatsQuery
 from .atlas import AtlasQuery
+from .colibri import ColibriQuery
 from .fink import FinkQuery
 from .gaia2dis import Gaia2Dis
 from .gaia_dr3 import GaiaDr3Query
@@ -18,10 +19,11 @@ from .vsx import VsxQuery
 from .ztf_periodic import ZtfPeriodicQuery
 
 ALERCE_QUERY = AlerceQuery("Alerce")
+ATLAS_QUERY = AtlasQuery("ATLAS")
 ANTARES_QUERY = AntaresQuery("Antares")
 ASTROCATS_QUERY = AstrocatsQuery("Astrocats")
+COLIBRI_QUERY = ColibriQuery("Astro-COLIBRI")
 FINK_QUERY = FinkQuery("Fink")
-ATLAS_QUERY = AtlasQuery("ATLAS")
 GAIA2_DIS = Gaia2Dis("Gaia DR2 Distances")
 GAIA_DR3 = GaiaDr3Query("Gaia DR3")
 GAIA_EDR3_DIS = GaiaEdr3Dis("Gaia EDR3 Distances")